LOG4J2-2253 Update ParameterConsumer index type from short to int
Originally ParameterVisitable was built for ReusableMessage which
provides a "short getParameterCount()" method. Now that the
interface isn't bound to ReusableMessage it might as well use
integer.

+++ log4j-api/src/main/java/org/apache/logging/log4j/message/ParameterConsumer.java
@@ -21,6 +21,6 @@ public interface ParameterConsumer<S> {
      * @param parameterIndex Index of the parameter
      * @param state
      */
-    void accept(Object parameter, short parameterIndex, S state);
+    void accept(Object parameter, int parameterIndex, S state);
 
 }
